Reader: remove Flux-powered follow button (#14468)
* Update README

* Remove Flux version of reader/follow-button

* Remove Flux version of blocks/follow-button

* repoint a file to the main follow-button
