+++ client/blocks/follow-button/index.jsx
@@ -1,71 +1,52 @@
 /**
  * External Dependencies
  */
-import React from 'react';
-import { noop } from 'lodash';
+import React, { Component } from 'react';
+import { noop, omitBy, isUndefined } from 'lodash';
+import { connect } from 'react-redux';
 
 /**
  * Internal Dependencies
  */
 import FollowButton from './button';
-import FeedSubscriptionActions from 'lib/reader-feed-subscriptions/actions';
-import FeedSubscriptionStore from 'lib/reader-feed-subscriptions';
+import { isFollowing } from 'state/selectors';
+import { follow, unfollow } from 'state/reader/follows/actions';
 
-const FollowButtonContainer = React.createClass( {
-	propTypes: {
+class FollowButtonContainer extends Component {
+	static propTypes = {
 		siteUrl: React.PropTypes.string.isRequired,
 		iconSize: React.PropTypes.number,
 		onFollowToggle: React.PropTypes.func,
 		followLabel: React.PropTypes.string,
 		followingLabel: React.PropTypes.string,
-	},
+		feedId: React.PropTypes.number,
+		siteId: React.PropTypes.number,
+	};
 
-	getDefaultProps() {
-		return {
-			onFollowToggle: noop,
-		};
-	},
+	static defaultProps = {
+		onFollowToggle: noop,
+	};
+	handleFollowToggle = following => {
+		if ( following ) {
+			const followData = omitBy(
+				{
+					feed_ID: this.props.feedId,
+					blog_ID: this.props.siteId,
+				},
+				isUndefined
+			);
 
-	getInitialState() {
-		return this.getStateFromStores();
-	},
-
-	getStateFromStores( props = this.props ) {
-		return { following: FeedSubscriptionStore.getIsFollowingBySiteUrl( props.siteUrl ) };
-	},
-
-	componentDidMount() {
-		FeedSubscriptionStore.on( 'change', this.onStoreChange );
-	},
-
-	componentWillUnmount() {
-		FeedSubscriptionStore.off( 'change', this.onStoreChange );
-	},
-
-	componentWillReceiveProps( nextProps ) {
-		this.updateState( nextProps );
-	},
-
-	updateState( props = this.props ) {
-		const newState = this.getStateFromStores( props );
-		if ( newState.following !== this.state.following ) {
-			this.setState( newState );
+			this.props.follow( this.props.siteUrl, followData );
+		} else {
+			this.props.unfollow( this.props.siteUrl );
 		}
-	},
-
-	onStoreChange() {
-		this.updateState();
-	},
-
-	handleFollowToggle( following ) {
-		FeedSubscriptionActions[ following ? 'follow' : 'unfollow' ]( this.props.siteUrl );
 		this.props.onFollowToggle( following );
-	},
+	};
 
 	render() {
 		return (
 			<FollowButton
-				following={ this.state.following }
+				following={ this.props.following }
 				onFollowToggle={ this.handleFollowToggle }
 				iconSize={ this.props.iconSize }
 				tagName={ this.props.tagName }
@@ -75,7 +56,15 @@ const FollowButtonContainer = React.createClass( {
 				className={ this.props.className }
 			/>
 		);
-	},
-} );
+	}
+}
 
-export default FollowButtonContainer;
+export default connect(
+	( state, ownProps ) => ( {
+		following: isFollowing( state, { feedUrl: ownProps.siteUrl } ),
+	} ),
+	{
+		follow,
+		unfollow,
+	}
+)( FollowButtonContainer );
